Move `InetNetwork` to Talpid package
