+++ android/src/main/kotlin/net/mullvad/mullvadvpn/model/TunConfig.kt
@@ -1,6 +1,7 @@
 package net.mullvad.mullvadvpn.model
 
 import java.net.InetAddress
+import net.mullvad.talpid.tun_provider.InetNetwork
 
 data class TunConfig(
     val addresses: List<InetAddress>,
