+++ mullvad-jni/src/lib.rs
@@ -43,7 +43,6 @@ const CLASSES_TO_LOAD: &[&str] = &[
     "net/mullvad/mullvadvpn/model/GetAccountDataResult$InvalidAccount",
     "net/mullvad/mullvadvpn/model/GetAccountDataResult$RpcError",
     "net/mullvad/mullvadvpn/model/GetAccountDataResult$OtherError",
-    "net/mullvad/mullvadvpn/model/InetNetwork",
     "net/mullvad/mullvadvpn/model/KeygenEvent$NewKey",
     "net/mullvad/mullvadvpn/model/KeygenEvent$Failure",
     "net/mullvad/mullvadvpn/model/KeygenFailure$TooManyKeys",
@@ -73,6 +72,7 @@ const CLASSES_TO_LOAD: &[&str] = &[
     "net/mullvad/talpid/net/TransportProtocol$Tcp",
     "net/mullvad/talpid/net/TransportProtocol$Udp",
     "net/mullvad/talpid/net/TunnelEndpoint",
+    "net/mullvad/talpid/tun_provider/InetNetwork",
     "net/mullvad/talpid/tunnel/ActionAfterDisconnect$Block",
     "net/mullvad/talpid/tunnel/ActionAfterDisconnect$Nothing",
     "net/mullvad/talpid/tunnel/ActionAfterDisconnect$Reconnect",
